Simplify platform related types
