+++ src/lib/dom.ts
@@ -4,30 +4,27 @@ import {
     SETTINGS_BTN_SELECTOR,
 } from "@/qualityMenuInjector";
 
-type ElementSelector =
-    | "video"
-    | typeof PANEL_MENU_SELECTOR
-    | typeof SETTINGS_BTN_SELECTOR
-    | typeof QUALITY_MENU_BTN_SELECTOR;
-
 type ElementSelectorMap = {
     video: HTMLVideoElement;
     [PANEL_MENU_SELECTOR]: HTMLDivElement;
     [SETTINGS_BTN_SELECTOR]: HTMLButtonElement;
     [QUALITY_MENU_BTN_SELECTOR]: HTMLDivElement;
 };
-export async function waitForElement<T extends ElementSelector>(
+
+export async function waitForElement<T extends keyof ElementSelectorMap>(
     selector: T,
     timeout: number = 10_000,
 ): Promise<ElementSelectorMap[T] | undefined> {
+    const findElement = () => document.querySelector<ElementSelectorMap[T]>(selector);
+
     return new Promise<ElementSelectorMap[T] | undefined>((resolve) => {
-        const element = document.querySelector<ElementSelectorMap[T]>(selector);
+        const element = findElement();
         if (element) {
             return resolve(element);
         }
 
         const observer = new MutationObserver(() => {
-            const element = document.querySelector<ElementSelectorMap[T]>(selector);
+            const element = findElement();
             if (element) {
                 observer.disconnect();
                 clearTimeout(timeoutId);
@@ -42,7 +39,7 @@ export async function waitForElement<T extends ElementSelector>(
 
         const timeoutId = setTimeout(() => {
             observer.disconnect();
-            return resolve(document.querySelector<ElementSelectorMap[T]>(selector) ?? undefined);
+            return resolve(findElement() ?? undefined);
         }, timeout);
     });
 }
