+++ src/lib/formatting.ts
@@ -1,5 +1,8 @@
 import { filesize } from "filesize";
 
+/**
+ * @example perHourDisplay(1_000_000) => "3600.00 MB/hour"
+ */
 export function perHourDisplay(sizePerSecondBytes: number): string {
     const sizePerHourMB = sizePerHour(sizePerSecondBytes);
     if (sizePerHourMB >= 1000) {
@@ -8,28 +11,31 @@ export function perHourDisplay(sizePerSecondBytes: number): string {
     return `${sizePerHourMB.toFixed(2)} MB/hour`;
 }
 
-export function totalSizeLiveDisplay(sizePerSecondBytes: number, durationSeconds?: number): string {
-    if (!durationSeconds) return "";
-    const sizePerMinuteMB = sizePerMinute(sizePerSecondBytes);
-    const totalSizeMB = (sizePerMinuteMB / 60) * durationSeconds;
-    if (totalSizeMB >= 1000) {
-        return `${(totalSizeMB / 1000).toFixed(2)} GB`;
-    }
-    return `${totalSizeMB.toFixed(2)} MB`;
-}
+/**
+ * @example totalSizeVideoDisplay(1_000_000) => "1.00 MB"
+ */
 export function totalSizeVideoDisplay(totalSizeBytes: number): string {
     return filesize(totalSizeBytes, { base: 10, standard: "jedec", round: 2 });
 }
 
+/**
+ * @example perMinuteDisplay(1_000_000) => "60.0 MB/min"
+ */
 export function perMinuteDisplay(sizePerSecondBytes: number): string {
     const sizePerMinuteMB = sizePerMinute(sizePerSecondBytes);
     return `${sizePerMinuteMB.toFixed(1)} MB/min`;
 }
 
+/**
+ * @example sizePerMinute(1_000_000) => 60.0
+ */
 export function sizePerMinute(sizePerSecondBytes: number): number {
     return (sizePerSecondBytes * 60) / 1_000_000;
 }
 
+/**
+ * @example sizePerHour(1_000_000) => 3600.0
+ */
 function sizePerHour(sizePerSecondBytes: number): number {
     return sizePerMinute(sizePerSecondBytes) * 60;
 }
