+++ src/pages/popup/header.tsx
@@ -1,5 +1,5 @@
 import type { PopupData } from "@app-types/uiTypes";
-import type { KickData, TwitchData, YoutubeData } from "@app-types/types";
+import type { KickData, TwitchData, YoutubeData } from "@app-types/platforms.types";
 import { humanizeDuration } from "@lib/utils";
 import { useNavigate } from "react-router";
