+++ src/qualityMenuInjector.ts
@@ -1,6 +1,6 @@
 /* eslint-disable unicorn/no-top-level-assignment-in-function */
 
-import type { YoutubeData } from "@app-types/types";
+import type { YoutubeData } from "@app-types/platforms.types";
 import { totalSizeVideoDisplay } from "@lib/formatting";
 import { waitForElement } from "@lib/dom";
