+++ src/resolution.ts
@@ -1,7 +1,7 @@
 /* eslint-disable unicorn/no-top-level-assignment-in-function */
 
 import { showTwitchToast, showYoutubeToast } from "@pages/toaster";
-import type { KickData, TwitchData, YoutubeData } from "@app-types/types";
+import type { KickData, TwitchData, YoutubeData } from "@app-types/platforms.types";
 import CONFIG from "@lib/constants";
 import { getFromSyncCache } from "@lib/cache";
