+++ src/tests/youtube.test.ts
@@ -1,4 +1,4 @@
-import type { ytInitialPlayerResponse, RawFormat } from "@app-types/types";
+import type { ytInitialPlayerResponse, RawFormat } from "@app-types/platforms.types";
 import fs from "node:fs";
 import path from "node:path";
 import {
