+++ src/types/types.ts
@@ -1,64 +1,4 @@
-import type { ytInitialSchema } from "@lib/schema";
-import { z } from "zod";
-
-export type ytInitialPlayerResponse = z.infer<typeof ytInitialSchema>;
-
-export type RawFormat = {
-    id: string;
-    title: string;
-    durationSeconds: number;
-    isLive: boolean;
-    formats: {
-        formatId: number;
-        height: number;
-        sizeBytes: number;
-        maxSizeBytes?: number;
-        bitrateBitsPerSecond?: number;
-    }[];
-    audioFormats: {
-        formatId: number;
-        sizeBytes: number;
-    }[];
-};
-
-type SuccessResponse<T> = {
-    success: true;
-    data: T;
-    createdAt?: string;
-};
-type ErrorResponse = {
-    success: false;
-    message: string;
-};
-
-export type YoutubeVideoFormat = {
-    formatId: number;
-    height: number;
-    sizeBytes: number;
-    maxSizeBytes?: number;
-    sizePerSecondBytes: number;
-};
-
-export type YoutubeVideoData = {
-    type: "video";
-    formats: YoutubeVideoFormat[];
-    title: string;
-    durationSeconds: number;
-    id: string;
-    isShorts?: boolean;
-    thumbnailUrl: string | undefined;
-    channelName: string | undefined;
-};
-
-type YoutubeLiveData = {
-    type: "live";
-    formats: StreamInfo[];
-    channelName: string;
-    thumbnailUrl: string | undefined;
-};
-
-export type YoutubeData = YoutubeVideoData | YoutubeLiveData;
-export type YoutubeBackgroundResponse = SuccessResponse<YoutubeData> | ErrorResponse;
+import type { KickData, TwitchData, YoutubeData } from "./platforms.types";
 
 export type StorageData<T extends YoutubeData | TwitchData | KickData> = {
     data: T;
@@ -73,28 +13,6 @@ export type OptionsMap = {
     cacheTTL?: number;
     qualityIds?: Record<string, boolean>;
     qualityMenu?: boolean;
-};
-
-export type StreamInfo = {
-    sizePerSecondBytes: number;
-    resolution: number;
-};
-
-export type TwitchLiveData = { type: "live"; data: StreamInfo[]; channelName: string };
-export type TwitchVodData = {
-    type: "vod";
-    data: StreamInfo[];
-    vodId: string;
-    durationSeconds: number | undefined;
-};
-
-export type TwitchData = TwitchLiveData | TwitchVodData;
-export type TwitchBackgroundResponse = SuccessResponse<TwitchData> | ErrorResponse;
-
-export type TwitchTokenData = {
-    value: string;
-    signature: string;
-    durationSeconds?: number;
 };
 
 export type FrontEndMessage =
@@ -136,21 +54,5 @@ export type KickVodMessage = {
     vodId: string;
     streamId: string;
 };
+
 export type KickMessage = KickLiveMessage | KickVodMessage;
-
-type KickLiveData = {
-    type: "live";
-    data: StreamInfo[];
-    channelName: string;
-};
-
-type KickVodData = {
-    type: "vod";
-    data: StreamInfo[];
-    vodId: string;
-    channelName: string | undefined;
-    durationSeconds: number | undefined;
-};
-
-export type KickData = KickLiveData | KickVodData;
-export type KickBackgroundResponse = SuccessResponse<KickData> | ErrorResponse;
